Refactor routing and layout: Header file
Refactor routing and layout: moved Header to App.js and conditionally hidden it on login route to avoid duplication across pages
