+++ admin-panel/src/App.js
@@ -1,11 +1,11 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
 import React from 'react';
-import { Route, Routes } from 'react-router-dom';
+import { Route, Routes, useLocation } from 'react-router-dom';
 
 import VcmAdminDashboard from './pages/VCM_Admin/VcmAdminDashboard';
 import RegisterCompanyAdmins from './pages/VCM_Admin/RegisterCompanyAdmins';
 import ViewCompanyAdmins from './pages/VCM_Admin/ViewCompanyAdmins';
-import AdminDashboard from './pages/Company_Admin/AdminDashboard';
+import AdminDashboard from './pages/AdminDashboard';
 import DeleteAdmin from './components/DeleteAdmin';
 import EditAdmin from './pages/Company_Admin/EditAdmins';
 import LoginPage from './pages/LoginPage';
@@ -16,11 +16,20 @@ import ViewAdmins from './pages/Company_Admin/ViewAdmins';
 import WarehouseGRNDashboard from './pages/WarehouseGRNDashboard';
 import WarehouseIssuingDashboard from './pages/WarehouseIssuingDashboard';
 import Settings from './components/Setting';
+import Header from './components/Header';
 
 
 function App() {
+  const location = useLocation();
+
+  // hide header on specific routes
+  const hideHeaderRoutes = ['/', '/login'];
+  
+  const hideHeader = hideHeaderRoutes.includes(location.pathname);
+
   return (
     <>
+      {!hideHeader && <Header />}
       <Routes>
         <Route path='/' element={<LoginPage />} />
         <Route path='/vcm-admin-dashboard/*' element={<VcmAdminDashboard />} />
