+++ admin-panel/src/pages/pdcDashboard.js
@@ -1,23 +1,7 @@
 import React from 'react';
-import Header from '../components/Header';
-import { Routes, Route } from 'react-router-dom';
 import { Container, Card } from 'react-bootstrap';
 
 export default function PdcDashboard() {
-  return (
-    <div className="dashboard-container">
-      <Header role='PDC' data-testid="header-toggle-button" />
-
-      <main className="dashboard-content">
-        <Routes>
-          <Route path="" element={<PdcHome />} />
-        </Routes>
-      </main>
-    </div>
-  );
-}
-
-function PdcHome() {
   return (
     <Container fluid>
       <Card className="dashboard-card">
